move funcionality into service
