+++ Adliance.Togglr/ExitCode.cs
@@ -1,0 +1,9 @@
+﻿namespace Adliance.Togglr
+{
+    public enum ExitCode
+    {
+        Ok = 0,
+        Error = -1,
+        Parameters = -2
+    }
+}
